+++ CHANGES.rst
@@ -5,6 +5,8 @@
 
   - Add ``:working-directory`` option to ``flycheck-define-command-checker``
     [GH-973] [GH-1012]
+  - ``flycheck-go-build-install-deps`` turns on dependency installation for ``go test``
+    as well as ``go build`` [GH-1003]
 
 - Improvements: